requirements.txt
# Core
pandas>=2.2.0,<3
numpy>=2.0.0,<3
scikit-learn>=1.4.0,<2
xgboost>=2.0.0,<3
sqlalchemy>=2.0.0,<3
psycopg2-binary>=2.9.9,<3

# MinIO / S3
minio>=7.2.0,<8
boto3>=1.34.0,<2

# Feast (local fork with OL mapper fix for PostgreSQL data source identity)
feast[postgres,redis] @ file:///app/wheels/feast-0.0.0-py3-none-any.whl

# MLflow
mlflow>=3.10.0,<4

# OpenLineage — official client (needed by Feast OL emitter)
openlineage-python>=1.0.0

# OpenLineage — client-side MLflow tracking store wrapper + KFP adapter
# Installed from local copy in Dockerfile (./openlineage-oai[mlflow])

# Kubeflow Pipelines
kfp>=2.7.0,<3

# Serving
fastapi>=0.110.0,<1
uvicorn>=0.29.0,<1
pydantic>=2.6.0,<3
httpx>=0.27.0,<1

# Validation (native OL action incompatible with GE 1.x; OL events emitted manually)
great-expectations>=1.0.0,<2

# Utilities
python-dotenv>=1.0.0,<2
PyYAML>=6.0,<7

# RAG / Vector DB
sentence-transformers>=2.2.0,<3
pymilvus>=2.3.0,<3
torch>=2.0.0,<3
